refactor: moved BssEpisode, BssSubject and FileType classes to Model folder and added  ParseBooleanStringToShort method
