+++ src/Shared/Model/FileType.cs
@@ -1,0 +1,7 @@
+namespace NHS.ServiceInsights.EpisodeIntegrationService;
+
+public enum FileType
+{
+    Episodes = 0,
+    Subjects = 1,
+}
